# specs/001-carte-arretes-rave-teknival/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: API Carte des arrêtés d'interdiction de rassemblements musicaux non déclarés
  description: >
    API publique en lecture seule exposant l'état des départements français
    (vert / rouge / gris) et l'historique complet des arrêtés préfectoraux
    d'interdiction des rassemblements musicaux non déclarés (rave party,
    teknival). Cet outil est un service d'information et ne remplace pas une
    vérification officielle auprès de la préfecture ou du Recueil des Actes
    Administratifs (RAA) concerné.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Version 1 de l'API (lecture seule, publique, sans authentification)

paths:
  /departements:
    get:
      operationId: getDepartementsAtDate
      summary: État de tous les départements à une date donnée
      description: >
        Retourne, pour chaque département, son état (vert/rouge/gris) à la
        date demandée, ainsi que la référence de l'arrêté applicable pour les
        départements en rouge et le connecteur associé pour les départements
        couverts.
      parameters:
        - name: date
          in: query
          required: true
          description: Date au format ISO 8601 (YYYY-MM-DD), interprétée en fuseau Europe/Paris.
          schema:
            type: string
            format: date
            example: "2026-08-10"
      responses:
        "200":
          description: État de tous les départements à la date demandée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartementsStateResponse"
        "400":
          description: Paramètre `date` manquant ou mal formé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /departements/{code}/evenements:
    get:
      operationId: getDepartementHistory
      summary: Historique complet des événements d'un département
      description: >
        Retourne la liste chronologique complète des événements (arrêtés,
        prolongations, levées) d'un département. Si le département n'a jamais
        été couvert par un connecteur, retourne une liste vide accompagnée du
        statut "non couvert".
      parameters:
        - name: code
          in: path
          required: true
          description: Code INSEE du département (ex. `75`, `2A`, `971`).
          schema:
            type: string
            example: "77"
      responses:
        "200":
          description: Historique du département (peut être vide si non couvert).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartementHistoryResponse"
        "404":
          description: Code de département inconnu (ne correspond à aucun département français).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /departements/etats:
    get:
      operationId: getDepartementsEtatsPeriode
      summary: États de tous les départements précalculés sur un intervalle (segments contigus)
      description: >
        Feature 006 (réglette sans appel réseau par pas) : retourne, pour
        chaque département, l'état déjà résolu par la même règle que
        `GET /departements?date=...`, sous forme de segments contigus non
        chevauchants couvrant intégralement `[debut, fin]`. `connecteur_id`
        et `derniere_collecte` sont portés une seule fois par département,
        jamais dupliqués par segment (ils ne dépendent pas du jour historique
        consulté). Le frontend charge cette route une seule fois à la
        sélection d'un intervalle, puis retrouve localement (comparaison de
        bornes) l'état à afficher à chaque pas de la réglette, sans dupliquer
        la règle métier (Principe 9).
      parameters:
        - name: debut
          in: query
          required: true
          description: Borne de début de l'intervalle (ISO 8601, YYYY-MM-DD, incluse).
          schema:
            type: string
            format: date
            example: "2026-05-01"
        - name: fin
          in: query
          required: true
          description: Borne de fin de l'intervalle (ISO 8601, YYYY-MM-DD, incluse).
          schema:
            type: string
            format: date
            example: "2026-05-31"
      responses:
        "200":
          description: États précalculés de tous les départements sur l'intervalle demandé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartementsEtatsPeriodeResponse"
        "400":
          description: >
            Paramètres `debut`/`fin` manquants, mal formés, `fin` antérieure
            à `debut`, ou intervalle dépassant la limite de taille autorisée.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /evenements:
    get:
      operationId: listEvenements
      summary: Tous les événements sur un intervalle de dates, tous départements confondus
      parameters:
        - name: debut
          in: query
          required: true
          description: Borne de début de l'intervalle (ISO 8601, YYYY-MM-DD, incluse).
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: fin
          in: query
          required: true
          description: Borne de fin de l'intervalle (ISO 8601, YYYY-MM-DD, incluse).
          schema:
            type: string
            format: date
            example: "2026-12-31"
      responses:
        "200":
          description: Liste des événements dont l'intervalle d'application recoupe la période demandée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  derniere_mise_a_jour:
                    type: string
                    format: date-time
                    description: Date de dernière mise à jour du jeu de données global (FR-012).
                  evenements:
                    type: array
                    items:
                      $ref: "#/components/schemas/Evenement"
        "400":
          description: Paramètres `debut`/`fin` manquants, mal formés, ou `fin` antérieure à `debut`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Etat:
      type: string
      enum: [vert, rouge, gris]
      description: >
        vert = aucun arrêté d'interdiction actif (département couvert) ;
        rouge = arrêté d'interdiction actif ; gris = aucun connecteur de
        collecte actif pour ce département à cette date (donnée non
        disponible, ne signifie pas "aucune interdiction").

    DepartementState:
      type: object
      required: [code, nom, etat]
      properties:
        code:
          type: string
          example: "77"
        nom:
          type: string
          example: "Seine-et-Marne"
        etat:
          $ref: "#/components/schemas/Etat"
        evenement_applicable:
          allOf:
            - $ref: "#/components/schemas/Evenement"
          nullable: true
          description: Présent uniquement quand `etat` est `rouge`.
        connecteur_id:
          type: string
          nullable: true
          description: Identifiant du connecteur ayant produit la donnée ; `null` si `etat` est `gris`.
        derniere_collecte:
          type: string
          format: date-time
          nullable: true
          description: >
            Date de dernière collecte réussie de la source pour ce département
            (feature 004, FR-001/FR-002). `null` si `etat` est `gris`, ou si le
            connecteur qui couvre ce département n'a encore jamais collecté
            avec succès.
        dernier_arrete_connu:
          allOf:
            - $ref: "#/components/schemas/DernierArreteConnu"
          nullable: true
          description: >
            Dernier arrêté d'interdiction/prolongation déjà terminé (idée n°2
            du backlog produit). Présent uniquement quand `etat` est `vert`
            et qu'un arrêté antérieur existe pour ce département ; `null`
            sinon (y compris pour un `vert` sans aucun historique).

    DernierArreteConnu:
      type: object
      required: [date_debut, date_fin]
      description: >
        Distinct de `Evenement` : `date_fin` est toujours renseignée ici (la
        date à afficher), résolue soit depuis la `date_fin` propre de
        l'arrêté, soit depuis la date de la levée qui y a mis fin quand
        `date_fin` est restée `null` sur l'événement stocké.
      properties:
        reference_arrete:
          type: string
          nullable: true
          example: "AP-2026-0501"
        date_debut:
          type: string
          format: date-time
        date_fin:
          type: string
          format: date-time

    DepartementsStateResponse:
      type: object
      required: [date, derniere_mise_a_jour, departements]
      properties:
        date:
          type: string
          format: date
          example: "2026-08-10"
        derniere_mise_a_jour:
          type: string
          format: date-time
          description: Date de dernière mise à jour du jeu de données global (FR-012).
        departements:
          type: array
          items:
            $ref: "#/components/schemas/DepartementState"

    DepartementStateSegment:
      type: object
      required: [date_debut, date_fin, etat]
      description: >
        Feature 006 : une période contiguë `[date_debut, date_fin]` (bornes
        incluses) pendant laquelle l'état du département reste inchangé.
        Mêmes champs d'état que `DepartementState` (hors `code`/`nom`/
        `connecteur_id`/`derniere_collecte`, portés une seule fois au niveau
        du département, cf. `DepartementEtatsPeriode`).
      properties:
        date_debut:
          type: string
          format: date
        date_fin:
          type: string
          format: date
        etat:
          $ref: "#/components/schemas/Etat"
        evenement_applicable:
          allOf:
            - $ref: "#/components/schemas/Evenement"
          nullable: true
          description: Présent uniquement quand `etat` est `rouge`.
        dernier_arrete_connu:
          allOf:
            - $ref: "#/components/schemas/DernierArreteConnu"
          nullable: true
          description: Présent uniquement quand `etat` est `vert` et qu'un arrêté antérieur existe.

    DepartementEtatsPeriode:
      type: object
      required: [code, nom, segments]
      properties:
        code:
          type: string
          example: "77"
        nom:
          type: string
          example: "Seine-et-Marne"
        connecteur_id:
          type: string
          nullable: true
        derniere_collecte:
          type: string
          format: date-time
          nullable: true
        segments:
          type: array
          items:
            $ref: "#/components/schemas/DepartementStateSegment"

    DepartementsEtatsPeriodeResponse:
      type: object
      required: [debut, fin, derniere_mise_a_jour, departements]
      properties:
        debut:
          type: string
          format: date
        fin:
          type: string
          format: date
        derniere_mise_a_jour:
          type: string
          format: date-time
        departements:
          type: array
          items:
            $ref: "#/components/schemas/DepartementEtatsPeriode"

    DepartementHistoryResponse:
      type: object
      required: [code, nom, couvert, evenements]
      properties:
        code:
          type: string
          example: "77"
        nom:
          type: string
          example: "Seine-et-Marne"
        couvert:
          type: boolean
          description: false si aucun connecteur n'a jamais couvert ce département ("non couvert").
        evenements:
          type: array
          items:
            $ref: "#/components/schemas/Evenement"

    Evenement:
      type: object
      required:
        - id
        - departement_code
        - type_evenement
        - date_debut
        - date_fin
        - date_saisie
        - connecteur_id
        - methode_collecte
      properties:
        id:
          type: string
          format: uuid
        departement_code:
          type: string
          example: "77"
        type_evenement:
          type: string
          enum: [interdiction, levee, prolongation]
        date_debut:
          type: string
          format: date-time
          description: ISO 8601 UTC.
        date_fin:
          type: string
          format: date-time
          nullable: true
          description: ISO 8601 UTC ; `null` = actif jusqu'à preuve du contraire.
        reference_arrete:
          type: string
          nullable: true
          example: "AP-2026-0842"
        source_url:
          type: string
          format: uri
          nullable: true
        date_saisie:
          type: string
          format: date-time
          description: Date d'entrée de l'événement dans le système (distincte de date_debut).
        connecteur_id:
          type: string
          example: "prefecture-77"
        methode_collecte:
          type: string
          enum: [automatique, manuelle_verifiee]

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "invalid_parameter"
        message:
          type: string
          example: "Le paramètre 'date' doit être au format YYYY-MM-DD."